+++ ctxtool/cancel.go
@@ -43,9 +43,12 @@ func (ac *AutoCancel) With(ctx context.Context, cancel context.CancelFunc) conte
 	return ctx
 }
 
-// FromCanceller creates a new context from a canceller. If a contex is passed,
-// then Deadline and Value will be ignored.
+// FromCanceller creates a new context from a canceller. If a value that
+// implements contex.Context is passed, then the value will be returned as is.
 func FromCanceller(c canceller) context.Context {
+	if ctx, ok := c.(context.Context); ok {
+		return ctx
+	}
 	return cancelContext{c}
 }